// backend/src/helpers/enums/HttpHeaderEnums.ts
export enum HttpHeaderEnums {
  PROXY_AUTHORIZATION = "proxy-authorization",
  TRANSACTION_ID = "x-transaction-id",
  ACCEPT_LANGUAGE = "accept-language",
  FORWARDED_FOR = "x-forwarded-for",
  USER_AGENT = "user-agent",
  ORIGIN = "origin",
  REFERRER = "referrer",

  CONTENT_TYPE = "content-type",
}