// backend/src/helpers/enums/HttpContentTypeEnums.ts
export enum HttpContentTypeEnums {
  APPLICATION_JSON = "application/json",
  APPLICATION_XML = "application/xml",
  APPLICATION_X_WWW_FORM_URLENCODED = "application/x-www-form-urlencoded",
  MULTIPART_FORM_DATA = "multipart/form-data",
  TEXT_HTML = "text/html",
  TEXT_PLAIN = "text/plain",
  TEXT_XML = "text/xml",
}